Improve the evaluation package structure
